Moved categoryControllerHelper to new folder 'repository'
