+++ controllers/overviewController.js
@@ -1,5 +1,12 @@
+const CategoryControllerHelper = require("../repository/CategoryRepository")
+
 function getIndexOverviewCategory(req, res) {
-    res.send("hello!")
+    const categoryControllerHelper = new CategoryControllerHelper();
+
+    categoryControllerHelper.searchCategory(req.params.category)
+        .then(result => console.log(result));
+
+    res.send(req.params.category)
  }
  
  module.exports = {
